+++ packages/react-components/src/card/index.js
@@ -1,9 +1,11 @@
 import Dialog from './components/dialog'
 import ArticleCard from './components/article-card'
+import ShortStory from './components/short-story'
 
-export { Dialog, ArticleCard }
+export { Dialog, ArticleCard, ShortStory }
 
 export default {
   Dialog,
   ArticleCard,
+  ShortStory,
 }