+++ packages/react-components/CHANGELOG.md
@@ -3,6 +3,92 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [8.25.0-rc.6](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/react-components@8.25.0-rc.5...@twreporter/react-components@8.25.0-rc.6) (2024-02-20)
+
+**Note:** Version bump only for package @twreporter/react-components
+
+
+
+
+
+# [8.25.0-rc.5](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/react-components@8.25.0-rc.4...@twreporter/react-components@8.25.0-rc.5) (2024-02-20)
+
+**Note:** Version bump only for package @twreporter/react-components
+
+
+
+
+
+# [8.25.0-rc.4](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/react-components@8.25.0-rc.3...@twreporter/react-components@8.25.0-rc.4) (2024-02-16)
+
+
+### Bug Fixes
+
+* footer styles ([79718d1](https://github.com/twreporter/twreporter-npm-packages/commit/79718d12552a18fd73c3d546d0cc210df153eaf0))
+
+
+
+
+
+# [8.25.0-rc.3](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/react-components@8.25.0-rc.2...@twreporter/react-components@8.25.0-rc.3) (2024-02-15)
+
+
+### Bug Fixes
+
+* export ShortStory ([e3f2c6d](https://github.com/twreporter/twreporter-npm-packages/commit/e3f2c6d8f6c0f03483abe4efbd173dd3ccc55b07))
+
+
+
+
+
+# [8.25.0-rc.2](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/react-components@8.25.0-rc.1...@twreporter/react-components@8.25.0-rc.2) (2024-02-15)
+
+**Note:** Version bump only for package @twreporter/react-components
+
+
+
+
+
+# [8.25.0-rc.1](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/react-components@8.25.0-rc.0...@twreporter/react-components@8.25.0-rc.1) (2024-02-07)
+
+
+### Features
+
+* add more PillButton theme color ([6e28d07](https://github.com/twreporter/twreporter-npm-packages/commit/6e28d079c9c54cd60d75c8fac453ff5d378aece9))
+* add more TextButton theme color & loading ([d72c845](https://github.com/twreporter/twreporter-npm-packages/commit/d72c845e4f73857ff10d6a0293d37402c0d29775))
+* iconWithTextButton P3 => P4 ([27caf83](https://github.com/twreporter/twreporter-npm-packages/commit/27caf832bf455be2d139aa55adfecc9996700596))
+
+
+
+
+
+# [8.25.0-rc.0](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/react-components@8.24.1-rc.1...@twreporter/react-components@8.25.0-rc.0) (2024-02-07)
+
+
+### Features
+
+* add `my reading` related entry on universal navigator ([79cadc2](https://github.com/twreporter/twreporter-npm-packages/commit/79cadc293493bc517fcf95139ff208633acd699b))
+
+
+
+
+
+## [8.24.1-rc.1](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/react-components@8.24.1-rc.0...@twreporter/react-components@8.24.1-rc.1) (2024-01-25)
+
+**Note:** Version bump only for package @twreporter/react-components
+
+
+
+
+
+## [8.24.1-rc.0](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/react-components@8.24.0...@twreporter/react-components@8.24.1-rc.0) (2024-01-24)
+
+**Note:** Version bump only for package @twreporter/react-components
+
+
+
+
+
 # [8.24.0](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/react-components@8.24.0-rc.6...@twreporter/react-components@8.24.0) (2024-01-24)
 
 **Note:** Version bump only for package @twreporter/react-components